feat(avatar): add custom avatar support

+++ src/utils/image/avatar.py
@@ -3,7 +3,7 @@ import collections
 from datetime import timedelta
 from io import BytesIO
 from pathlib import Path
-from typing import Deque
+from typing import Deque, Literal
 
 from aiohttp import ClientSession, ClientTimeout
 from async_lru import alru_cache
@@ -44,6 +44,7 @@ class Avatar:
     LOCAL_FALLBACK_DIR = "data/dynamic/avatar"
     GROUP_PATH = "group/{id}.png"
     USER_PATH = "user/{id}.png"
+    USER_CUSTOM_PATH = "user-custom/{id}.jpg"
 
     avatar_timeout: float
     avatar_concurrency: int
@@ -52,6 +53,13 @@ class Avatar:
 
     _queue_handled = False
     _queue: Deque[tuple[int, bool]] = collections.deque()
+
+    @classmethod
+    def session(cls) -> ClientSession:
+        if cls._session is None:
+            cls._session = ClientSession(timeout=ClientTimeout(
+                total=cls.avatar_timeout))
+        return cls._session
 
     @staticmethod
     @alru_cache(ttl=timedelta(minutes=5).total_seconds())
@@ -87,17 +95,27 @@ class Avatar:
         is_group: bool = False,
         default: Image.Image | None = None,
     ) -> Image.Image | None:
-        if cls._session is None:
-            cls._session = ClientSession(timeout=ClientTimeout(
-                total=cls.avatar_timeout))
-        return await cls._fetch_with_local(cls._session, id,
+        return await cls._fetch_with_local(cls.session(), id,
                                            is_group) or default
+
+    @classmethod
+    def load_custom(cls, user_id: int) -> Image.Image | None:
+        file = Path(
+            cls.LOCAL_FALLBACK_DIR) / cls.USER_CUSTOM_PATH.format(id=user_id)
+        if file.exists():
+            try:
+                return Image.open(file)
+            except Exception:
+                file.unlink()
+        return None
 
     @classmethod
     async def user(cls,
                    user_id: int,
                    default: Image.Image | None = None) -> Image.Image | None:
         await cls._start_queue()
+        if avatar := cls.load_custom(user_id):
+            return avatar
         return await Avatar.fetch(user_id, is_group=False, default=default)
 
     @classmethod
@@ -108,12 +126,35 @@ class Avatar:
         return await Avatar.fetch(group_id, is_group=True, default=default)
 
     @classmethod
-    def clear_user_local(cls, user_id: int) -> bool:
-        file = Path(cls.LOCAL_FALLBACK_DIR) / cls.USER_PATH.format(id=user_id)
-        if file.exists():
-            file.unlink()
-            return True
-        return False
+    async def update_user_avatar(
+        cls,
+        user_id: int,
+        avatar: str | Image.Image | None,
+    ) -> Literal["updated", "cache", "custom", "none"]:
+        if avatar is None:
+            # remove both
+            for template in (cls.USER_PATH, cls.USER_CUSTOM_PATH):
+                file = Path(
+                    cls.LOCAL_FALLBACK_DIR) / template.format(id=user_id)
+                if file.exists():
+                    file.unlink()
+                    return "cache" if template == cls.USER_PATH else "custom"
+            return "none"
+        else:
+            file = Path(cls.LOCAL_FALLBACK_DIR) / cls.USER_CUSTOM_PATH.format(
+                id=user_id)
+            file.parent.mkdir(parents=True, exist_ok=True)
+            if isinstance(avatar, str):
+                avatar = await cls._fetch(cls.session(), avatar)
+            dim = min(avatar.size)
+            l = (avatar.width - dim) // 2
+            t = (avatar.height - dim) // 2
+            r = l + dim
+            b = t + dim
+            cropped = avatar.crop((l, t, r, b))
+            cropped.thumbnail((640, 640))
+            cropped.convert("RGB").save(file)
+            return "updated"
 
     @classmethod
     async def _start_queue(cls):
